chg: dev: added check in install when deleting directories

+++ install.sh
@@ -1,6 +1,6 @@
 pip3 uninstall nmtf -y || echo "No nmtf to uninstall"
 pip3 install setuptools
 python3 setup.py install
-rm -r dist
-rm -r build
-rm -r nmtf.egg-info*
+if [ -d "dist" ] ; then rm -r dist ; fi
+if [ -d "build" ] ; then rm -r build ; fi
+if ls nmtf.egg-info* &> /dev/null ; then rm -r nmtf.egg-info* ; fi
